refactor: rename test files and fix imports
